Revert "moving api calls from patientTable to patients, first working version of the search tool (search by name with exact and start with)"
This reverts commit 5679b5d79b1289cf7e1678dd75841da2f3da4aff.
