+++ src/components/patients/patientTable/index.tsx
@@ -3,16 +3,15 @@ import { Link } from "react-router-dom";
 import { Cell, Column, Table } from "@blueprintjs/table";
 import { Icon, H3 } from "@blueprintjs/core";
 import { ROUTE_PATIENT } from "../../../constants";
+import { getPatients, getCount } from "services/api";
 import { Patient } from "types";
 
 import "./style.css";
 
-interface Props {
-  patients: Patient[];
-  patientCount: string;
-}
+const PatientTable = () => {
+  const [patients, setPatients] = React.useState([] as Patient[]);
+  const [patientCount, setPatientCount] = React.useState("");
 
-const PatientTable = ({ patients, patientCount }: Props) => {
   const renderPatientAttribute = (
     attribute: "id" | "identifier" | "firstName" | "lastName" | "age",
     index: number
@@ -25,6 +24,16 @@ const PatientTable = ({ patients, patientCount }: Props) => {
       </React.Fragment>
     </Cell>
   );
+
+  React.useEffect(() => {
+    const fetchPatients = async () => {
+      const patients: Patient[] = await getPatients();
+      setPatients(patients);
+      const count = await getCount("Patient", { _summary: "count" });
+      setPatientCount(count);
+    };
+    fetchPatients();
+  }, []);
 
   return (
     <>
@@ -75,7 +84,7 @@ const PatientTable = ({ patients, patientCount }: Props) => {
         </Table>
       </div>
       <div className="infoPatient">
-        {patientCount !== undefined && `${patientCount} patients identifiés`}
+        {patientCount && `${patientCount} patients identifiés`}
       </div>
     </>
   );
